fix: 회원관리 | JWT token seq int to long 변경 [S07P31D205-46]

+++ Backend/sdutyplus/src/main/java/com/d205/sdutyplus/domain/warn/controller/WarnController.java
@@ -37,7 +37,7 @@ public class WarnController {
     })
     @PostMapping("/user/{toUserSeq}")
     public ResponseEntity<ResponseDto> userWarn(@ApiIgnore Authentication auth, @PathVariable Long toUserSeq) {
-        long fromUserSeq = (int)auth.getPrincipal();
+        Long fromUserSeq = (Long)auth.getPrincipal();
 
         final boolean success = warnService.userWarn(fromUserSeq, toUserSeq);
         if (success) {
